+++ __init__.py
@@ -4,7 +4,7 @@ bl_info = {
     "name": "Storytools - Storyboard Tools",
     "description": "Set of tools for Storyboarding",
     "author": "Samuel Bernou",
-    "version": (0, 9, 4),
+    "version": (1, 0, 0),
     "blender": (3, 3, 0),
     "location": "View3D",
     "warning": "",
@@ -18,6 +18,7 @@ from . import properties
 from . import preferences
 from . import OP_story_palettes
 from . import OP_camera_controls
+from . import OP_camera_data
 from . import OP_gp_objects
 from . import GZ_toolbar
 from . import OP_workspace_setup
@@ -34,6 +35,7 @@ modules = (
     OP_story_palettes,
     OP_gp_objects,
     OP_camera_controls,
+    OP_camera_data,
     handles,
     GZ_toolbar,
     OP_workspace_setup,